Created .app from installer
I tried the copy_from_dmg, and it has serious downsides. Either I have to explicitly create an installs array, which will be incomplete or tied to a specific release (which would defeat the point of the recipe) or I leave it up to munkiimport to create an installs array, in which case it creates an installs array based on the Scratch 2 installer and not the actual launcher .app. As far as I can tell, the simplest way to get Munki to automatically create the correct installs array is to just copy the contained folder to its on .app and then have Munki import that.